+++ datafusion/catalog-listing/Cargo.toml
@@ -53,6 +53,7 @@ object_store = { workspace = true }
 url = { workspace = true }
 
 [dev-dependencies]
+async-trait = { workspace = true }
 tempfile = { workspace = true }
 tokio = { workspace = true }
